Refactorin aggregation report stuff into different files
