+++ python/aggregation/report.py
@@ -1,128 +1,13 @@
-from typing import Dict
 import pandas as pd
 from os import path as osp
 from io import StringIO
-from collections import namedtuple
 import matplotlib.pyplot as plt
 from matplotlib.lines import Line2D
 import mplcursors
 import webbrowser
 import tempfile
 import pydoc
-
-TEST_TREE = r"C:\PioSOLVER\Reports\SimpleTree\SRP\b25\BTNvBB"
-
-
-ranks = {
-    "2": 2,
-    "3": 3,
-    "4": 4,
-    "5": 5,
-    "6": 6,
-    "7": 7,
-    "8": 8,
-    "9": 9,
-    "T": 10,
-    "J": 11,
-    "Q": 12,
-    "K": 13,
-    "A": 14,
-}
-
-
-ranks_rev = {
-    2: "2",
-    3: "3",
-    4: "4",
-    5: "5",
-    6: "6",
-    7: "7",
-    8: "8",
-    9: "9",
-    10: "T",
-    11: "J",
-    12: "Q",
-    13: "K",
-    14: "A",
-}
-
-
-def ahml(rank):
-    if rank == 14:
-        return "A"
-    if rank > 9:
-        return "H"
-    if rank > 5:
-        return "M"
-    else:
-        return "L"
-
-
-def color_texture(texture):
-    """
-    Return a coloration of a texture
-    """
-
-    red = "00"
-    green = "00"
-    blue = "00"
-
-    if "MONOTONE" in texture:
-        red = "ff"
-    elif "FD" in texture:
-        red = "99"
-    elif "RAINBOW" in texture:
-        red = "00"
-    else:
-        print(f"Warning: Unrecognized suitedness {texture}")
-
-    if "STRAIGHT" in texture:
-        green = "ff"
-    elif "OESD" in texture:
-        green = "aa"
-    elif "GUTSHOT" in texture:
-        green = "55"
-    elif "DISCONNECTED" in texture:
-        green = "00"
-    else:
-        print(f"Warning: Unrecognized connectedness {texture}")
-
-    if "TOAK" in texture:
-        blue = "ff"
-    elif "PAIRED" in texture:
-        blue = "99"
-    elif "UNPAIRED" in texture:
-        blue = "00"
-    else:
-        print(f"Warning: Unrecognized pairedness {texture}")
-
-    return f"#{red}{green}{blue}"
-
-
-def marker_size_from_high_card(flop, max_size=None, min_size=10):
-    if max_size is None:
-        max_size = 220
-    r, s = card_tuple(flop.split()[0])
-    factor = (max_size / min_size) ** (1 / 12)
-    size = min_size * factor ** (r - 2)
-    return size
-
-
-def card_tuple(c):
-    r, s = c.strip()
-    return ranks[r], s
-
-
-Info = namedtuple("Info", ["player", "node_id", "line", "starting_stacks"])
-
-
-def parse_info(info: str):
-    lines = info.splitlines()
-    player = lines[0].strip().split()[0]
-    node_id = lines[1][9:].strip()
-    line = lines[3][6:].strip()
-    starting_stacks = int(lines[4].split(":")[1].strip())
-    return Info(player, node_id, line, starting_stacks)
+from aggregation.util import *
 
 
 class AggregationReport:
@@ -687,86 +572,3 @@ class AggregationReport:
 
     def paginate(self):
         pydoc.pager(self.dump())
-
-
-class TreeNode:
-    def __init__(self, path):
-        """
-        Args:
-            path (str): file path to directory corresponding to node in agg tree
-        """
-        self.path = path
-        self._children: Dict[str, TreeNode] = {}
-        self._info = None
-        self._report = None
-
-    def __getitem__(self, item):
-        return self._children.get(item, default=None)
-
-    def get_info(self) -> str:
-        """
-        Get node info, if it exists. Otherwise return the empty string
-        """
-        if self._info is None:
-            info_path = osp.join(self.path, "info.txt")
-            self._info = ""
-            if osp.exists(info_path):
-                with open(info_path, "r") as f:
-                    self._info = f.read().strip()
-        return self._info
-
-    def get_report(self) -> AggregationReport:
-        if self._report is None:
-            if osp.exists(self.path):
-                self._report = AggregationReport(self.path)
-        return self._report
-
-
-class AggregationTreeViewer:
-    """
-    View and navigate a tree of aggregation reports.
-
-    The base of the tree contains the Root directory, which corresponds to the
-    root of the PIO tree.
-
-    ```
-    base/
-      +-- Root/
-            +-- Check/
-            |      +-- Bet 11/
-            |      |     +-- Call/
-            |      |     +-- Fold/
-            |      +-- Check/     # To the turn
-            +-- Bet 11/
-                  +-- Call/
-                  +-- Fold/
-                  +-- Raise/
-                        +-- .../
-    ```
-    """
-
-    def __init__(self, base: str):
-        self.base = base
-        self._path = []
-
-
-def get_test_report_r_x():
-    r = osp.join(TEST_TREE, "Root")
-    r_x = osp.join(r, "CHECK")
-    assert osp.isdir(r_x)
-
-    report_path = osp.join(r_x, "report.csv")
-    assert osp.isfile(report_path)
-    agg_report = AggregationReport(r_x)
-    return agg_report
-
-
-def test_aggregation_report():
-    print("Testing aggregation reports")
-    agg_report = get_test_report_r_x()
-    print(agg_report._df)
-    print("Done testing aggregation reports")
-
-
-if __name__ == "__main__":
-    test_aggregation_report()
